early code for supporting both gocardless APIs
i hoped there would be more backwards compatibility with the old API in
the new API but that seems to not be the case. i don't want to create a
new distribution for this as i can reuse a lot of the Client code (and
probably some classes too)

the plan is to have Business::GoCardless::Basic now extend the top level
Business::GoCardless class, and have Business::GoCardless::Pro support
the new API. we can do some checking in the B::G::Client class to using
the api_version to check/complain as necessary. once the Basic API is
gone we can remove all the Basic API related code and just have the Pro
version be the default (we then flip the logic on its head the the top
level namespace will extend B::G::Pro)

it's looking like B::G::Pro is *not* going to be a drop in replacement
for B::G::Basic, although i might attempt to make that possible at the
start