fix #13 - always have Content-Length header on POST/PUT
even when there is no content - this seems to be enforced by something
in Google App Engine as the original IETF RFCs[1] state that the
Content-Length SHOULD be included (i.e. not MUST be included) and
GoCardless have recently moved to Google Cloud[2] and Google have
tightened up their requirements so now POST/PUT requests with no
Content-Length are receiving a "Error 411 (Length Required)!!1" error

arguably this is a bug in whatever Google are doing - lack of a
Content-Length could be taken as zero content length (or, i suppose, an
unknown length), as per the original RFCs, and it probably shouldn't
break a shit load of client code that was working before

this could possibly be pushed down as a patch into LWP, but there is the
possibility that doing *that* will cause other problems because :shrug:
software

bump VERSION and Changes for CPAN release

[1] https://tools.ietf.org/html/rfc7230#section-3.3.2

[2] "That's correct, we did move to Google Cloud earlier this year. I'm
sorry that this caused an issue with your integration, but I'm glad that
you managed to resolve the problem.

    Apologies for the lack of a heads up, unfortunately this is
something that was overlooked as it seems most clients send the header
through by default.

    – GoCardless Support"